pyproject.toml
[project]
name = "agentx"
version = "0.1.0"
description = "Green Agent (Assessor) for Custom MCP Task Evaluation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.0",
    "mcp>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.110.0",
    "pytest>=8.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "langchain-mcp-adapters>=0.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
